# hello-goa/gen/http/openapi3.yaml
openapi: 3.0.3
info:
    title: Goa API
    version: 0.0.1
servers:
    - url: http://localhost:80
      description: Default server for hello
paths:
    /greetHello/{name}:
        get:
            tags:
                - hello
            summary: greetHello hello
            description: To get hello
            operationId: hello#greetHello
            parameters:
                - name: name
                  in: path
                  description: Name to greet
                  required: true
                  schema:
                    type: string
                    description: Name to greet
                    example: Perspiciatis laboriosam voluptatem sunt maxime quae.
                  example: Et molestiae.
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                type: string
                                description: A greeting message
                                example: Et cum autem ipsa laboriosam.
                            example: Porro quas nobis ea.
    /respondHello/{name}:
        get:
            tags:
                - hello
            summary: respondToHello hello
            description: To respond to hello
            operationId: hello#respondToHello
            parameters:
                - name: name
                  in: path
                  description: Name to greet
                  required: true
                  schema:
                    type: string
                    description: Name to greet
                    example: Quae est possimus eius architecto.
                  example: Accusamus dicta blanditiis velit totam hic.
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                type: string
                                description: A response message
                                example: Reprehenderit aut quae ad voluptas.
                            example: Quas dolor id assumenda nobis qui est.
components: {}
tags:
    - name: hello
      description: A simple service that says hello